+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/HttpGoogleServicesDAO.scala
@@ -82,7 +82,7 @@ class HttpGoogleServicesDAO(
   subEmail: String,
   pemFile: String,
   appsDomain: String,
-  orgID: Int,
+  orgID: Long,
   groupsPrefix: String,
   appName: String,
   deletedBucketCheckSeconds: Int,
